requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
qdrant-client==1.13.1
sentence-transformers==3.4.1
httpx==0.27.2
pypdf==5.2.0
python-multipart==0.0.9
langgraph==0.2.55
langchain-core==0.3.40
sentry-sdk==2.19.2
prometheus-client==0.20.0
python-dateutil==2.9.0.post0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg[binary]==3.1.19
redis==5.0.4
rq==1.16.2
